+++ truffle/src/com.oracle.truffle.api.interop.java/src/com/oracle/truffle/api/interop/java/JavaInterop.java
@@ -280,7 +280,12 @@ public final class JavaInterop {
      * @since 0.9
      */
     public static TruffleObject asTruffleObject(Object obj) {
-        return asTruffleObject(obj, currentPolyglotContext());
+        // legacy behavior: treat class as static class
+        return asTruffleObject(obj, currentPolyglotContext(), true);
+    }
+
+    static TruffleObject asTruffleObject(Object obj, Object languageContext) {
+        return asTruffleObject(obj, languageContext, false);
     }
 
     /**
@@ -289,11 +294,15 @@ public final class JavaInterop {
      * @param obj a Java object to convert into one suitable for <em>Truffle</em> languages
      * @return converted object
      */
-    static TruffleObject asTruffleObject(Object obj, Object languageContext) {
+    static TruffleObject asTruffleObject(Object obj, Object languageContext, boolean asStaticClass) {
         if (obj instanceof TruffleObject) {
             return ((TruffleObject) obj);
         } else if (obj instanceof Class) {
-            return JavaObject.forClass((Class<?>) obj, languageContext);
+            if (asStaticClass) {
+                return JavaObject.forStaticClass((Class<?>) obj, languageContext);
+            } else {
+                return JavaObject.forClass((Class<?>) obj, languageContext);
+            }
         } else if (obj == null) {
             return JavaObject.NULL;
         } else if (obj.getClass().isArray()) {