+++ truffle/src/com.oracle.truffle.api.interop.java/src/com/oracle/truffle/api/interop/java/JavaInteropAccessor.java
@@ -67,6 +67,11 @@ final class JavaInteropAccessor extends Accessor {
                 }
                 return ((JavaFunctionObject) object).getDescription();
             }
+
+            @Override
+            public Object asStaticClassObject(Class<?> clazz, Object hostLanguageContext) {
+                return JavaObject.forStaticClass(clazz, hostLanguageContext);
+            }
         };
     }