+++ truffle/src/com.oracle.truffle.api.interop.java/src/com/oracle/truffle/api/interop/java/JavaInteropReflect.java
@@ -32,11 +32,11 @@ import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.lang.reflect.Proxy;
 import java.lang.reflect.Type;
-import java.util.Collection;
-import java.util.LinkedHashSet;
 import java.util.Objects;
 import java.util.function.BiFunction;
 import java.util.function.Supplier;
+
+import org.graalvm.collections.EconomicSet;
 
 import com.oracle.truffle.api.CallTarget;
 import com.oracle.truffle.api.CompilerAsserts;
@@ -152,8 +152,13 @@ final class JavaInteropReflect {
             }
         }
 
-        if (!readable) {
-            if (onlyStatic) {
+        if (onlyStatic) {
+            if (!readable) {
+                if ("class".equals(name)) {
+                    readable = true;
+                }
+            }
+            if (!readable) {
                 Class<?> innerClass = findInnerClass(clazz, name);
                 if (innerClass != null) {
                     readable = true;
@@ -234,12 +239,13 @@ final class JavaInteropReflect {
     }
 
     @CompilerDirectives.TruffleBoundary
-    static String[] findUniquePublicMemberNames(Class<?> clazz, boolean onlyInstance, boolean includeInternal) throws SecurityException {
+    static String[] findUniquePublicMemberNames(Class<?> clazz, boolean onlyStatic, boolean includeInternal) throws SecurityException {
         JavaClassDesc classDesc = JavaClassDesc.forClass(clazz);
-        Collection<String> names = new LinkedHashSet<>();
-        names.addAll(classDesc.getFieldNames(!onlyInstance));
-        names.addAll(classDesc.getMethodNames(!onlyInstance, includeInternal));
-        if (!onlyInstance) {
+        EconomicSet<String> names = EconomicSet.create();
+        names.addAll(classDesc.getFieldNames(onlyStatic));
+        names.addAll(classDesc.getMethodNames(onlyStatic, includeInternal));
+        if (onlyStatic) {
+            names.add("class");
             if (Modifier.isPublic(clazz.getModifiers())) {
                 // no support for non-static member types now
                 for (Class<?> t : clazz.getClasses()) {
@@ -250,7 +256,7 @@ final class JavaInteropReflect {
                 }
             }
         }
-        return names.toArray(new String[0]);
+        return names.toArray(new String[names.size()]);
     }
 
     @SuppressWarnings({"unchecked"})